mobile: abstract layout box component for maker configuration
slightly improve ui
