+++ mobile-app/src/components/app-maker/index.ts
@@ -1,3 +1,3 @@
 export { AnimatedFlingContainer } from './AnimatedFlingContainer';
 export { LayoutDivider } from './LayoutDivider';
-export { MakerConfiguration } from './MakerConfiguration';
+export { MakerConfiguration } from './maker-configuration';
