+++ mobile-app/src/components/app-maker/maker-configuration/index.ts
@@ -1,0 +1,1 @@
+export { MakerConfiguration } from './MakerConfiguration';
